+++ src/LightOps.Commerce.Services.Product.Backends.InMemory/Domain/QueryHandlers/FetchProductsByIdsQueryHandler.cs
@@ -8,20 +8,20 @@ using LightOps.Commerce.Services.Product.Backends.InMemory.Api.Providers;
 
 namespace LightOps.Commerce.Services.Product.Backends.InMemory.Domain.QueryHandlers
 {
-    public class FetchProductsByHandleQueryHandler : IFetchProductsByHandleQueryHandler
+    public class FetchProductsByIdsQueryHandler : IFetchProductsByIdsQueryHandler
     {
         private readonly IInMemoryProductProvider _inMemoryProductProvider;
 
-        public FetchProductsByHandleQueryHandler(IInMemoryProductProvider inMemoryProductProvider)
+        public FetchProductsByIdsQueryHandler(IInMemoryProductProvider inMemoryProductProvider)
         {
             _inMemoryProductProvider = inMemoryProductProvider;
         }
 
-        public Task<IList<IProduct>> HandleAsync(FetchProductsByHandleQuery query)
+        public Task<IList<IProduct>> HandleAsync(FetchProductsByIdsQuery query)
         {
             var products = _inMemoryProductProvider
                 .Products
-                .Where(c => query.Handles.Contains(c.Handle))
+                .Where(c => query.Ids.Contains(c.Id))
                 .ToList();
 
             return Task.FromResult<IList<IProduct>>(products);