+++ src/LightOps.Commerce.Services.Product/Api/Queries/FetchProductsByHandlesQuery.cs
@@ -3,7 +3,7 @@ using LightOps.CQRS.Api.Queries;
 
 namespace LightOps.Commerce.Services.Product.Api.Queries
 {
-    public class FetchProductsByHandleQuery : IQuery
+    public class FetchProductsByHandlesQuery : IQuery
     {
         public IList<string> Handles { get; set; }
     }