+++ src/LightOps.Commerce.Services.Product/Api/Queries/FetchProductsByIdsQuery.cs
@@ -3,7 +3,7 @@ using LightOps.CQRS.Api.Queries;
 
 namespace LightOps.Commerce.Services.Product.Api.Queries
 {
-    public class FetchProductsByIdQuery : IQuery
+    public class FetchProductsByIdsQuery : IQuery
     {
         public IList<string> Ids { get; set; }
     }